Embedded StockItem entity in InventoryItem, modified tests.
