+++ src/test/java/com/um/inventoryservice/DataLayer/StockItemRepositoryTest.java
@@ -16,9 +16,6 @@ class StockItemRepositoryTest {
     @Autowired
     StockItemRepository stockItemRepository;
 
-    @Autowired
-    InventoryItemRepository inventoryItemRepository;
-
     StockItem stockItem = buildStockItem();
 
     String STOCK_ID = stockItem.getStockItemId();
@@ -114,20 +111,6 @@ class StockItemRepositoryTest {
                 .verifyComplete();
     }
 
-    @Test
-    public void insertInventoryItem (){
-        InventoryItem inventoryItem = buildInventoryItem();
-
-        Publisher<InventoryItem> setup = inventoryItemRepository.deleteAll().thenMany(inventoryItemRepository.save(inventoryItem));
-        StepVerifier
-                .create(setup)
-                .consumeNextWith(foundItem -> {
-                    assertEquals(inventoryItem.getInventoryItemId(), foundItem.getInventoryItemId());
-                    assertEquals(inventoryItem.getStockItemId(), foundItem.getStockItemId());
-                    assertEquals(inventoryItem.getQuantityInStock(), foundItem.getQuantityInStock());
-                })
-                .verifyComplete();
-    }
 
     private StockItem buildStockItem() {
         return StockItem.builder()
@@ -149,19 +132,5 @@ class StockItemRepositoryTest {
                 .build();
     }
 
-    private InventoryItem buildInventoryItem() {
-        return InventoryItem.builder()
-                .inventoryItemId("1176543")
-                .stockItemId("2212345")
-                .quantityInStock(265)
-                .build();
-    }
 
-    private InventoryItemDTO buildInventoryItemDTO() {
-        return InventoryItemDTO.builder()
-                .inventoryItemId("1123456")
-                .stockItemId("2213579")
-                .quantityInStock(75)
-                .build();
-    }
 }
